Further fixes for S3 methods declared with setMethodS3
1) automatically generates \method{}{} markup which should be correct if
and only if the generic name is everything up to but not including the
first period in the function name. Thus this will not work properly if a
generic function has a name like "my.generic" so that the specific function
is "my.generic.my_type".

2) fix for behaviour of package.skeleton, which, as used within this
package, for a function "[[.Distribution" generates a file with name
"z[[.Distribution.Rd"

3) prints S3method calls suitable for inclusion in a NAMESPACE file

The mechanism depends on the parse of the source file picking up and
setting the "created" field to "setMethodS3". Something similar might work
for other ways of creating S3 generics, but I am not currently clear how to
pick those up reliably.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/inlinedocs@84 edb9625f-4e0d-4859-8d74-9fd3b1da38cb